requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
nltk==3.8.1
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
django-environ==0.11.2 